Move classes to a more suitable namespace
