+++ Sources/Todo.Commons/Constants/Logging.cs
@@ -1,0 +1,23 @@
+﻿namespace Todo.Commons.Constants
+{
+    /// <summary>
+    /// Contains constants related to logging various events generated by this application and its host.
+    /// </summary>
+    public static class Logging
+    {
+        /// <summary>
+        /// Represents the name of the folder where log files will be created.
+        /// </summary>
+        public const string LogsHomeEnvironmentVariable = "LOGS_HOME";
+
+        /// <summary>
+        /// Represents the identifier of the conversation grouping several events related to the same operation.
+        /// </summary>
+        public const string ConversationId = "ConversationId";
+
+        /// <summary>
+        /// Represents the name of an application flow.
+        /// </summary>
+        public const string ApplicationFlowName = "ApplicationFlowName";
+    }
+}
