+++ Sources/Todo.DependencyInjection/ServicesModule.cs
@@ -2,8 +2,6 @@ namespace Todo.DependencyInjection
 {
     using Autofac;
 
-    using Commons;
-
     using Services.Security;
     using Services.TodoItemManagement;
 
@@ -19,14 +17,16 @@ namespace Todo.DependencyInjection
 
         protected override void Load(ContainerBuilder builder)
         {
-            bool isDevelopmentEnvironment = Constants.EnvironmentNames.Development.Equals(EnvironmentName);
-            bool isIntegrationTestsEnvironment = Constants.EnvironmentNames.IntegrationTests.Equals(EnvironmentName);
+            bool isDevelopmentEnvironment = Commons.Constants.EnvironmentNames.Development.Equals(EnvironmentName);
+
+            bool isIntegrationTestsEnvironment =
+                Commons.Constants.EnvironmentNames.IntegrationTests.Equals(EnvironmentName);
 
             var persistenceModule = new PersistenceModule
             {
                 ConnectionStringName = isIntegrationTestsEnvironment
-                    ? Constants.ConnectionStrings.UsedByIntegrationTests
-                    : Constants.ConnectionStrings.UsedByApplication,
+                    ? Commons.Constants.ConnectionStrings.UsedByIntegrationTests
+                    : Commons.Constants.ConnectionStrings.UsedByApplication,
                 EnableDetailedErrors = isDevelopmentEnvironment || isIntegrationTestsEnvironment,
                 EnableSensitiveDataLogging = isDevelopmentEnvironment || isIntegrationTestsEnvironment
             };
