+++ Sources/Todo.Logging/Serilog/SerilogActivator.cs
@@ -5,8 +5,6 @@ namespace Todo.Logging.Serilog
     using System.IO;
     using System.Linq;
 
-    using Commons;
-
     using global::Serilog;
 
     using Microsoft.Extensions.Configuration;
@@ -66,7 +64,7 @@ namespace Todo.Logging.Serilog
                 return;
             }
 
-            string logsHomeEnvironmentVariableName = Constants.Logging.LogsHomeEnvironmentVariable;
+            string logsHomeEnvironmentVariableName = Commons.Constants.Logging.LogsHomeEnvironmentVariable;
             string logsHomeDirectoryPath = Environment.GetEnvironmentVariable(logsHomeEnvironmentVariableName);
 
             if (string.IsNullOrWhiteSpace(logsHomeDirectoryPath) || !Directory.Exists(logsHomeDirectoryPath))
