+++ Tests/IntegrationTests/Todo.Persistence.IntegrationTests/TodoDbContextTests.cs
@@ -3,8 +3,6 @@ namespace Todo.Persistence
     using System;
     using System.Threading.Tasks;
 
-    using Commons;
-
     using Entities;
 
     using FluentAssertions;
@@ -147,13 +145,13 @@ namespace Todo.Persistence
             var configurationBuilder = new ConfigurationBuilder();
 
             IConfigurationRoot testConfiguration = configurationBuilder.AddJsonFile("appsettings.json", false)
-                .AddJsonFile($"appsettings.{Constants.EnvironmentNames.IntegrationTests}.json", false)
+                .AddJsonFile($"appsettings.{Commons.Constants.EnvironmentNames.IntegrationTests}.json", false)
                 .AddEnvironmentVariables()
                 .Build();
 
             // ReSharper disable once SettingNotFoundInConfiguration
             var testConnectionString =
-                testConfiguration.GetConnectionString(Constants.ConnectionStrings.UsedByIntegrationTests);
+                testConfiguration.GetConnectionString(Commons.Constants.ConnectionStrings.UsedByIntegrationTests);
 
             var connectionStringBuilder = new NpgsqlConnectionStringBuilder(testConnectionString)
             {
